backend/requirements.txt
flask
flask-cors
python-dotenv
requests
boto3
aws-wsgi